# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hospitality Ops Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues/{venueId}/temperature/equipment:
    get:
      operationId: listTemperatureEquipment
      tags: [temperature]
      summary: List temperature-controlled equipment for a venue
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TemperatureEquipment"
    post:
      operationId: createTemperatureEquipment
      tags: [temperature]
      summary: Create temperature-controlled equipment (admin only)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemperatureEquipmentInput"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureEquipment"
  /venues/{venueId}/temperature/equipment/{equipmentId}:
    patch:
      operationId: updateTemperatureEquipment
      tags: [temperature]
      summary: Update temperature equipment (admin only)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemperatureEquipmentInput"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureEquipment"
    delete:
      operationId: archiveTemperatureEquipment
      tags: [temperature]
      summary: Archive temperature equipment (admin only)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Archived
  /venues/{venueId}/temperature/summary:
    get:
      operationId: getTemperatureSummary
      tags: [temperature]
      summary: Get temperature summary for dashboard
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Temperature summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureSummary"
  /venues/{venueId}/temperature/logs:
    get:
      operationId: listTemperatureLogs
      tags: [temperature]
      summary: List temperature logs
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: equipmentId
          in: query
          schema:
            type: integer
        - name: logType
          in: query
          schema:
            type: string
        - name: failedOnly
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of temperature logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TemperatureLog"
    post:
      operationId: createTemperatureLog
      tags: [temperature]
      summary: Create a temperature log entry
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemperatureLogInput"
      responses:
        "201":
          description: Created log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureLog"
  /venues/{venueId}/temperature/logs/{logId}:
    patch:
      operationId: updateTemperatureLog
      tags: [temperature]
      summary: Update a temperature log (corrective action)
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemperatureLogUpdate"
      responses:
        "200":
          description: Updated log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureLog"
...
components:
  schemas:
...
    TemperatureEquipment:
      type: object
      required: [id, venueId, name, type, minTemp, maxTemp, isArchived, createdAt, updatedAt, isOverdue]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        name:
          type: string
        type:
          type: string
        minTemp:
          type: number
        maxTemp:
          type: number
        checkIntervalHours:
          type: integer
          nullable: true
        isArchived:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        lastChecked:
          type: string
          nullable: true
        lastTemp:
          type: number
          nullable: true
        lastStatus:
          type: string
          nullable: true
        isOverdue:
          type: boolean
        nextDueAt:
          type: string
          nullable: true
    TemperatureEquipmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        type:
          type: string
        minTemp:
          type: number
        maxTemp:
          type: number
        checkIntervalHours:
          type: integer
          nullable: true
    TemperatureLog:
      type: object
      required: [id, venueId, logType, recordedTemp, status, checkedBy, checkedAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        equipmentId:
          type: integer
          nullable: true
        invoiceId:
          type: integer
          nullable: true
        logType:
          type: string
        itemName:
          type: string
          nullable: true
        recordedTemp:
          type: number
        status:
          type: string
        notes:
          type: string
          nullable: true
        correctiveAction:
          type: string
          nullable: true
        recheckTemp:
          type: number
          nullable: true
        isResolved:
          type: boolean
          nullable: true
        checkedBy:
          type: string
        checkedAt:
          type: string
    TemperatureLogInput:
      type: object
      required: [recordedTemp, checkedBy]
      properties:
        equipmentId:
          type: integer
          nullable: true
        invoiceId:
          type: integer
          nullable: true
        logType:
          type: string
        itemName:
          type: string
          nullable: true
        recordedTemp:
          type: number
        notes:
          type: string
          nullable: true
        correctiveAction:
          type: string
          nullable: true
        checkedBy:
          type: string
    TemperatureLogUpdate:
      type: object
      properties:
        correctiveAction:
          type: string
          nullable: true
        recheckTemp:
          type: number
          nullable: true
        isResolved:
          type: boolean
          nullable: true
        notes:
          type: string
          nullable: true
    TemperatureEquipmentStatus:
      type: object
      required: [id, name, type, minTemp, maxTemp, isOverdue]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        minTemp:
          type: number
        maxTemp:
          type: number
        checkIntervalHours:
          type: integer
          nullable: true
        lastChecked:
          type: string
          nullable: true
        lastTemp:
          type: number
          nullable: true
        lastStatus:
          type: string
          nullable: true
        isOverdue:
          type: boolean
        nextDueAt:
          type: string
          nullable: true
    TemperatureSummary:
      type: object
      required: [failedToday, unresolvedFails, checksToday, equipmentCount, overdueCount, equipmentStatus]
      properties:
        failedToday:
          type: integer
        unresolvedFails:
          type: integer
        checksToday:
          type: integer
        equipmentCount:
          type: integer
        overdueCount:
          type: integer
        equipmentStatus:
          type: array
          items:
            $ref: "#/components/schemas/TemperatureEquipmentStatus"

    # ─── CHEMICAL ──────────────────────────────────────────────────────────────
...